Split things up to make dependency injection work without side effects
