+++ packages/apps/api-authorizer/package.json
@@ -7,6 +7,7 @@
   "license": "MIT",
   "scripts": {
     "build": "tsc --build tsconfig.build.json",
+    "test": "jest --silent --rootDir .",
     "clean": "rimraf ./lib tsconfig.build.tsbuildinfo"
   },
   "devDependencies": {
